requirements.txt
annotated-types==0.7.0
anyio==4.6.0
blessed==1.20.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
google-auth==2.35.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
kubernetes==30.1.0
llm-sandbox==0.1.1
oauthlib==3.2.2
openai==1.50.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
